+++ api/schemas/service_provider.py
@@ -7,20 +7,20 @@ from pydantic import BaseModel
 class ServiceProviderBase(BaseModel):
     id: int
     user_id: int
-    rating: int
+    rating: float
     title: str
     biography: str
 
 class ServiceProviderCreate(BaseModel):
     user_id: int
-    rating: int
+    rating: float
     title: str
     biography: str
 
 
 class ServiceProviderUpdate(BaseModel):
     user_id: Optional[int] = None
-    rating: Optional[int] = None
+    rating: Optional[float] = None
     title: Optional[str] = None
     biography: Optional[str] = None
